requirements.txt
streamlit== 1.42.0
PyPDF2==3.0.1
openai==1.63.0
python-dotenv==1.0.1